+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 Changes are grouped by the date they are merged to the main branch of the repository and are ordered from newest to oldest. Dates use the ISO 8601 extended calendar date format, i.e. YYYY-MM-DD.
 
+## 2024-12-13
+
+- Update to Rocket League harness for artifact management and new menu navigation.
+
 ## 2024-12-04
 
 - Update some timings in RDR2 harness.